Squashed commit of the following:

commit aac9aac7fdda1f5dd7ca37ac32690700156655eb
Author: Serhii Vlasiuk <svlasiuk@cisco.com>
Date:   Tue Mar 9 15:11:08 2021 +0200

    ips_options: add sticky buffer script_data ips option within normalized javascripts payload

    Update max value for js_normalization_depth = {-1, max53}
    Add mutual exclusion behaviour for js_normalization_depth and normalize_javascript
    js_normalization_depth - enables enhanced normalizer
    normalize_javascript - enables legacy normalizer

+++ src/utils/js_normalizer.cc
@@ -30,7 +30,7 @@
 using namespace snort;
 
 int JSNormalizer::normalize(const char* srcbuf, uint16_t srclen, char* dstbuf, uint16_t dstlen,
-        const char** ptr, int* bytes_copied, int norm_depth)
+        const char** ptr, int* bytes_copied, int64_t norm_depth)
 {
     std::stringstream in, out;
