+++ src/utils/js_normalizer.h
@@ -28,7 +28,7 @@ class JSNormalizer
 {
 public:
     static int normalize(const char* srcbuf, uint16_t srclen, char* dstbuf, uint16_t dstlen,
-        const char** ptr, int* bytes_copied, int norm_depth);
+        const char** ptr, int* bytes_copied, int64_t norm_depth);
 };
 }
